fix types

+++ matcha_core/src/generator.rs
@@ -21,7 +21,7 @@ impl MatchaGenerator {
         Ok(Self { session })
     }
 
-    pub fn synthesise(&self, symbols: Vec<i64>, scale: Scale) -> Result<(Array3<f32>, i32)> {
+    pub fn synthesise(&self, symbols: Vec<i64>, scale: Scale) -> Result<(Array3<f32>, i64)> {
         let symbols = intersperse(symbols, 0);
         let x = Array2::from_shape_vec((1, symbols.len()), symbols.clone())?;
         let x_lengths = arr1(&[symbols.len() as i64]);
@@ -31,7 +31,7 @@ impl MatchaGenerator {
             "scales" => scale.to_ndarray(),
         ]?)?;
         let audio_array = outputs["mel"].try_extract_tensor::<f32>()?.to_owned();
-        let mel_lengths = outputs["mel_lengths"].try_extract_tensor::<i32>()?;
+        let mel_lengths = outputs["mel_lengths"].try_extract_tensor::<i64>()?;
 
         Ok((
             Array3::from_shape_vec(
